containers/django/project/requirements.txt
Django>=2.0
Markdown>=2.1.0
django-filter>=1.0.1
django-guardian>=1.1.1
djangorestframework
djangorestframework-jwt
django-cors-headers>=2.4.0